@@ -8,7 +8,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -16,15 +16,15 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  intl: ^0.17.0
-  url_launcher: ^6.0.17
-  provider: ^6.0.1
-  get_it: ^7.2.0
+  intl: ^0.18.0
+  url_launcher: ^6.1.12
+  provider: ^6.0.5
+  get_it: ^7.6.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.17
-  lint: ^1.8.1
+  mockito: ^5.4.2
+  lint: ^2.1.2
 
 
 # For information on the generic Dart part of this file, see the
